backend/translation-service/requirements.txt
# Service de traduction Meeshy - Dépendances Python optimisées

# Framework web
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic-settings==2.0.3

# Communication gRPC et ZeroMQ
grpcio>=1.59.3
grpcio-tools>=1.59.3
grpcio-reflection>=1.59.0
protobuf>=4.25.1
pyzmq==25.1.1

# Machine Learning et NLP (versions compatibles)
torch>=2.0.0
transformers>=4.35.0
tokenizers>=0.15.0
accelerate>=0.24.0
datasets>=2.14.0
sentencepiece>=0.1.99

# Cache et base de données
redis==5.0.1
prisma==0.11.0

# Détection de langue et analyse de texte
langdetect==1.0.9
nltk>=3.8.1
textstat>=0.7.3

# Utilitaires système
psutil>=5.9.6
numpy>=1.24.4
pandas>=2.1.3
python-dotenv>=1.0.0
pydantic>=2.5.0
python-multipart>=0.0.6
aiofiles>=23.2.1

# Logging et monitoring
loguru>=0.7.2
prometheus-client==0.19.0

# Tests
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov==4.1.0
httpx>=0.25.2
